+++ packages/ui-core/package.json
@@ -31,6 +31,7 @@
         "@terascope/utils": "^0.10.1",
         "apollo-boost": "^0.3.1",
         "apollo-client": "^2.5.1",
+        "date-fns": "^1.30.1",
         "graphql": "^14.3.0",
         "prop-types": "^15.7.2",
         "query-string": "^6.5.0",
