+++ packages/ui-core/src/components/data-access/users/Create.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Page } from '../../core';
-import UserForm from '../UserForm';
+import UserForm from './UserForm';
 
-const CreateUser: React.FC = () => {
+const Create: React.FC = () => {
     return (
         <Page title="Create Users">
             <UserForm />
@@ -10,4 +10,4 @@ const CreateUser: React.FC = () => {
     );
 };
 
-export default CreateUser;
+export default Create;
