+++ packages/ui-core/src/components/data-access/users/Edit.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Page, tsWithRouter } from '../../core';
-import UserForm from '../UserForm';
+import UserForm from './UserForm';
 
 export default tsWithRouter(({ match }) => {
     return (
